backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
azure-identity==1.19.0
azure-storage-blob==12.23.1
pyodbc==5.2.0
sqlalchemy==2.0.36
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.6.1
